fix(mcpdoc): update readme cli example and mcp json. (#22)
See[ this issue
thread](https://github.com/langchain-ai/mcpdoc/issues/21).

Update examples in the readme to match the arg parser in `cli.py` which
is configured to append multiple urls:

```python
    parser.add_argument(
        "--urls",
        "-u",
        type=str,
        nargs="+",
        help="List of llms.txt URLs or file paths with optional names (format: 'url_or_path' or 'name:url_or_path')",
    )
```

The current examples in the readme file uses multiple `--url` flags so
the previous url is overridden with each new url flag.